+++ packages/observation-extractors/bitcoin-observation-extractor/lib/DogeEsploraObservationExtractor.ts
@@ -3,7 +3,7 @@ import { DogeEsploraRosenExtractor } from '@rosen-bridge/rosen-extractor';
 import { TokenMap } from '@rosen-bridge/tokens';
 import { AbstractLogger } from '@rosen-bridge/abstract-logger';
 import { DataSource } from 'typeorm';
-import { BitcoinEsploraTransaction } from '@rosen-bridge/bitcoin-esplora-scanner';
+import { BitcoinEsploraTransaction } from '@rosen-bridge/bitcoin-scanner';
 
 export class DogeEsploraObservationExtractor extends AbstractObservationExtractor<BitcoinEsploraTransaction> {
   readonly FROM_CHAIN = 'doge';