+++ packages/scanners/cardano-scanner/lib/interfaces/index.ts
@@ -1,0 +1,21 @@
+import { DataSource } from 'typeorm';
+
+interface OgmiosReconnectionConfig {
+  initialDelay?: number;
+  maxDelay?: number;
+  maxAttempts?: number;
+}
+
+interface CardanoOgmiosConfig {
+  nodeHostOrIp: string;
+  nodePort: number;
+  initialSlot: number;
+  initialHash: string;
+  maxTryBlock?: number;
+  dataSource: DataSource;
+  useTls?: boolean;
+  reconnectionConfig?: OgmiosReconnectionConfig;
+  suffix?: string;
+}
+
+export { OgmiosReconnectionConfig, CardanoOgmiosConfig };