+++ packages/scanners/cardano/.lintstagedrc.js
@@ -1,5 +1,0 @@
-export default {
-  '*.ts': () => 'npm run type-check',
-  '*.{js,ts}': ['eslint --fix', 'vitest related --run'],
-  '*': 'prettier --ignore-unknown --write',
-};