+++ packages/scanners/ergo-scanner/lib/index.ts
@@ -1,0 +1,3 @@
+export * from './ergoScanner';
+export * from './network/ergoNodeNetwork';
+export * from './network/ergoExplorerNetwork';